move util class
